add node_enhancement class and modify test
