+++ src/pai_rag/config/settings.toml
@@ -32,7 +32,7 @@ embed_batch_size = 10
 [rag.evaluation]
 retrieval = ["mrr", "hit_rate"]
 response_label = true
-response = ["Faithfulness", "Answer Relevancy", "Guideline Adherence", "Correctness", "Semantic Similarity"]
+response = ["Faithfulness", "Answer Relevancy", "Correctness", "Semantic Similarity"]
 
 [rag.index]
 persist_path = "localdata/storage"
@@ -49,6 +49,12 @@ name = "qwen-turbo"
 
 [rag.llm_chat_engine]
 type = "SimpleChatEngine"
+
+[rag.node_enhancement]
+enable_raptor = true
+max_token_in_cluster = 3000
+max_clusters = 52
+proba_threshold = 0.10
 
 [rag.node_parser]
 type = "Sentence"
