+++ src/pai_rag/modules/module_registry.py
@@ -24,6 +24,7 @@ MODULE_CONFIG_KEY_MAP = {
     "OssCacheModule": "cache",
     "EvaluationModule": "evaluation",
     "BM25IndexModule": "bm25",
+    "NodesEnhancementModule": "node_enhancement",
 }
